+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Orion Core
 
 This is the changelog for ```orionjs:core```.
 
+### 0.5.0
+
+- Reorder dictionary code.
+
+- Subscribe to dictionary on startup automatically.
+
 ### 0.4.10
 
 - Fixed bug #21.